refactor: 명확하게 Event를 명시하도록 변경
