+++ src/main/java/com/yoyomo/infra/aws/dto/UploadEvent.java
@@ -1,0 +1,4 @@
+package com.yoyomo.infra.aws.dto;
+
+public record UploadEvent(String subDomain) {
+}
