+++ src/main/java/com/yoyomo/infra/aws/s3/service/UploadEventListener.java
@@ -1,7 +1,7 @@
-package com.yoyomo.infra.redis;
+package com.yoyomo.infra.aws.s3.service;
 
 import com.yoyomo.domain.club.exception.UnavailableSubdomainException;
-import com.yoyomo.infra.aws.s3.service.S3Service;
+import com.yoyomo.infra.aws.dto.UploadEvent;
 import com.yoyomo.infra.redis.service.RedisQueueService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 @EnableAsync
-public class RedisQueueWorker {
+public class UploadEventListener {
     private static final String DOMAIN_FORMAT = "%s.crayon.land";
 
     private final S3Service s3Service;
@@ -20,13 +20,12 @@ public class RedisQueueWorker {
 
     @Async
     @EventListener
-    public void processUpload(String subDomain) {
-        String fullSubDomain = String.format(DOMAIN_FORMAT, subDomain);
+    public void processUpload(UploadEvent uploadEvent) {
+        String fullSubDomain = String.format(DOMAIN_FORMAT, uploadEvent.subDomain());
         try {
             s3Service.upload(fullSubDomain);
         } catch (Exception e) {
             retryUpload(fullSubDomain);
-            throw new UnavailableSubdomainException(e.getMessage());
         }
     }
